to_i

+++ app/models/restaurant_wheel.rb
@@ -4,7 +4,7 @@ class RestaurantWheel
   delegate :each_with_index, to: :wheel
 
   def initialize(num_stops: DEFAULT_NUM_STOPS)
-    @num_stops = num_stops.to_d
+    @num_stops = num_stops.to_i
   end
 
   def wheel
